+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+* Add `restore_symlink`
+* Strict imports for `upyog.os.utils`
+
+
 ## 0.7.10 -- 26 Jun 2024
 * Bugfixes
 * `image_to_base64` util to convert PIL -> base64 strings